requirements.txt
astroid==2.0.4
colorama==0.3.9
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
numpy==1.15.0
opencv-python==3.4.2.17
Pillow==5.2.0
pylint==2.1.1
six==1.11.0
wrapt==1.10.11
